requirements.txt
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
Flask-WTF==0.14.3
gitdb==4.0.7
GitPython==3.1.17
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.0
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.17.1
six==1.16.0
smmap==4.0.0
toml==0.10.2
Werkzeug==2.0.1
WTForms==2.3.3
